requirements.txt
requests>=2.31.0
supabase>=2.3.0
rapidfuzz>=3.6.1
python-dotenv>=1.0.0
schedule>=1.2.0
playwright>=1.40.0 